feat: change to fewer files and restructure saving method
