+++ src/scripts/make_phoneme_covering.py
@@ -1,14 +1,14 @@
 """Make a phoneme covering set of wikipedia articles.
 
 Usage:
-    python src/scripts/make_wiki_phoneme_covering.py sort_by=da
+    python src/scripts/make_phoneme_covering.py sort_by=da
 """
 
 import hydra
 from omegaconf import DictConfig
+from pathlib import Path
 
-from tts_text.make_phoneme_covering_set import make_phoneme_covering_set
-from tts_text.sort_wiki_by_phoneme_occurence import sort_wiki_by_phoneme_occurence
+from tts_text.phoneme_covering_set import make_phoneme_covering_set
 
 
 @hydra.main(config_path="../../config", config_name="config", version_base=None)
@@ -17,8 +17,8 @@ def main(config: DictConfig) -> None:
     Args:
         config: The Hydra config for your project.
     """
-    sort_wiki_by_phoneme_occurence(cfg=config)
-    make_phoneme_covering_set(cfg=config)
+    output_path = Path(config.dirs.data) / "raw"
+    make_phoneme_covering_set(cfg=config, output_path=output_path)
 
 
 if __name__ == "__main__":
